Fix for real-time validation of blocks within a page

+++ resources/assets/js/components/EditBlock.js
@@ -1,5 +1,11 @@
+/**
+ * the block editing form for editing a single block within a page
+ * displayed in the sidebar
+ *
+ * listens to validation events from block-form and updates validation
+ * within the the global vuex state
+ */
 import { mapState, mapMutations, mapGetters } from 'vuex';
-
 import EditOptions from 'components/EditOptions';
 
 export default {
@@ -7,6 +13,10 @@ export default {
 	name: 'edit-block',
 
 	extends: EditOptions,
+
+	created() {
+		eventBus.$on('block:setValidation', this.setValidation);
+	},
 
 	computed: {
 		...mapGetters([
@@ -53,15 +63,19 @@ export default {
 			'deleteBlockValidationIssue'
 		]),
 
+		/**
+		 * toggles the validation of this current block in the vuex state
+		 * @param {bool} status
+		 */
 		setValidation(status) {
-			if(status) {
-				this.deleteBlockValidationIssue(this.currentBlock.id);
-			}
-			else {
-				this.addBlockValidationIssue(this.currentBlock.id);
+			if (this.currentBlock) {
+				if (status) {
+					this.deleteBlockValidationIssue(this.currentBlock.id);
+				}
+				else {
+					this.addBlockValidationIssue(this.currentBlock.id);
+				}
 			}
 		}
-
 	}
-
 };
